fix an unlikely panic caused by external program taking too long to start

+++ misc/sys_unix.go
@@ -59,7 +59,7 @@ func InvokeProgram(envVars []string, timeoutSec int, program string, args ...str
 	var timedOut bool
 	timeOutTimer := time.AfterFunc(time.Duration(timeoutSec)*time.Second, func() {
 		timedOut = true
-		if !KillProcess(proc.Process) {
+		if proc.Process != nil && !KillProcess(proc.Process) {
 			logger.Warning("InvokeProgram", program, nil, "failed to kill after time limit exceeded")
 		}
 	})
@@ -74,6 +74,9 @@ func InvokeProgram(envVars []string, timeoutSec int, program string, args ...str
 
 // KillProcess kills the process or the group of processes associated with it.
 func KillProcess(proc *os.Process) (success bool) {
+	if proc == nil {
+		return true
+	}
 	// Kill process group if it is one
 	if killErr := syscall.Kill(-proc.Pid, syscall.SIGKILL); killErr == nil {
 		success = true
